+++ packages/@overeng/ci-tools/tsconfig.json.genie.ts
@@ -13,5 +13,5 @@ export default tsconfigJson({
     types: ['node', 'bun'],
   },
   include: ['src/**/*', 'bin/**/*.ts'],
-  references: [{ path: '../otel-contract' }, { path: '../utils' }, { path: '../utils-dev' }],
+  references: [{ path: '../otel-contract' }, { path: '../utils' }],
 } satisfies TSConfigArgs)