+++ packages/@overeng/genie/tsconfig.json.genie.ts
@@ -14,5 +14,5 @@ export default tsconfigJson({
     jsx: 'react-jsx',
   },
   include: ['src/**/*.ts', 'src/**/*.tsx', 'bin/**/*.ts', 'bin/**/*.tsx'],
-  references: [{ path: '../otel-contract' }, { path: '../utils' }, { path: '../utils-dev' }],
+  references: [{ path: '../otel-contract' }, { path: '../utils' }],
 } satisfies TSConfigArgs)