+++ packages/@overeng/notion-property-write/tsconfig.json.genie.ts
@@ -12,5 +12,5 @@ export default tsconfigJson({
     ...nodeTypes,
   },
   include: ['src/**/*'],
-  references: [{ path: '../notion-effect-schema' }, { path: '../utils-dev' }],
+  references: [{ path: '../notion-effect-schema' }],
 } satisfies TSConfigArgs)