+++ packages/@overeng/notion-react/tsconfig.json.genie.ts
@@ -23,6 +23,5 @@ export default tsconfigJson({
     { path: '../notion-md' },
     { path: '../otel-contract' },
     { path: '../utils' },
-    { path: '../utils-dev' },
   ],
 } satisfies TSConfigArgs)